Changed the shortcut of the add-on's submenu from D to W.
Added a technique to fix the problem when switching to shorter years, for instance, when switching from a common year to a leap year.

Since the calculation function executes a down arrow and an up arrow to find the value of the current year, it sometimes switches to shorter years, which changes the value of the day field in February month.